Refactor exclusive gateway tests
